requirements.txt
# Kafka client
kafka-python==2.0.2

# Web framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Redis client
redis==5.0.3

# Data / Math
numpy==1.26.4

# Testing
pytest==8.1.1
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
psutil==5.9.8
